backend/functions/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
python-dotenv
passlib[bcrypt]
pyjwt
email-validator
cryptography
flask-cors
